Move utils functions to proper folder
