+++ src/app/content/highlights/hooks/locationChange.ts
@@ -4,7 +4,7 @@ import { AppServices, MiddlewareAPI } from '../../../types';
 import { bookAndPage } from '../../selectors';
 import { receiveHighlights } from '../actions';
 import * as select from '../selectors';
-import addCurrentPageToSummaryFilters from './addCurrentPageToSummaryFilters';
+import { addCurrentPageToSummaryFilters } from '../utils';
 
 const hookBody = (services: MiddlewareAPI & AppServices) => async() => {
   const {dispatch, getState, highlightClient} = services;
