+++ src/app/content/highlights/utils/index.ts
@@ -1,0 +1,9 @@
+export { default as addCurrentPageToSummaryFilters } from './addCurrentPageToSummaryFilters';
+export { default as getHighlightLocationForPage } from './getHighlightLocationForPage';
+export { default as getHighlightLocationsFromBook } from './getHighlightLocationsFromBook';
+export {
+  addSummaryHighlight,
+  removeSummaryHighlight,
+  updateSummaryHighlight,
+  updateSummaryHighlightsDependOnFilters,
+} from './summaryHighlightsUtils';
