+++ src/app/content/highlights/utils/summaryHighlightsUtils.spec.ts
